sources: refactor rpm source into module.
Extract the Rpm source into `sources/_rpm.py`.

LP: #1648906
