+++ snapcraft/tests/sources/test_sources.py
@@ -16,12 +16,10 @@
 
 import copy
 import os
-import shutil
 import tarfile
 import unittest.mock
 
 import fixtures
-import libarchive
 
 from snapcraft.internal import common, sources
 from snapcraft import tests
@@ -205,47 +203,6 @@ class TestDeb(tests.FakeFileHTTPServerBasedTestCase):
 
         with open(deb_download, 'r') as deb_file:
             self.assertEqual('Test fake compressed file', deb_file.read())
-
-
-class TestRpm(tests.TestCase):
-
-    def test_pull_rpm_file_must_extract(self):
-        rpm_file_name = 'test.rpm'
-        dest_dir = 'src'
-        os.makedirs(dest_dir)
-
-        test_file_path = os.path.join(self.path, 'test.txt')
-        open(test_file_path, 'w').close()
-        rpm_file_path = os.path.join(self.path, rpm_file_name)
-        os.chdir(self.path)
-        with libarchive.file_writer(rpm_file_path, 'cpio', 'gzip') as rpm:
-            rpm.add_files('test.txt')
-
-        rpm_source = sources.Rpm(rpm_file_path, dest_dir)
-        rpm_source.pull()
-
-        self.assertEqual(os.listdir(dest_dir), ['test.txt'])
-
-    def test_extract_and_keep_rpmfile(self):
-        rpm_file_name = 'test.rpm'
-        dest_dir = 'src'
-        os.makedirs(dest_dir)
-
-        test_file_path = os.path.join(self.path, 'test.txt')
-        open(test_file_path, 'w').close()
-        rpm_file_path = os.path.join(self.path, rpm_file_name)
-        os.chdir(self.path)
-        with libarchive.file_writer(rpm_file_path, 'cpio', 'gzip') as rpm:
-            rpm.add_files('test.txt')
-
-        rpm_source = sources.Rpm(rpm_file_path, dest_dir)
-        # This is the first step done by pull. We don't call pull to call the
-        # second step with a different keep_rpm value.
-        shutil.copy2(rpm_source.source, rpm_source.source_dir)
-        rpm_source.provision(dst=dest_dir, keep_rpm=True)
-
-        test_output_files = ['test.txt', rpm_file_name]
-        self.assertCountEqual(os.listdir(dest_dir), test_output_files)
 
 
 class TestBazaar(SourceTestCase):
